[SMELL CODE] renamed packages
