+++ src/main/java/polytech/annecy/groupe_valentin/discreteBehaviorSimulator/ClockObserver.java
@@ -1,4 +1,4 @@
-package polytech.annecy.groupeValentin.discreteBehaviorSimulator;
+package polytech.annecy.groupe_valentin.discreteBehaviorSimulator;
 
 public interface ClockObserver {
 	public void clockChange(int time);
