+++ src/main/java/polytech/annecy/groupe_valentin/timer/DateTimer.java
@@ -1,4 +1,4 @@
-package polytech.annecy.groupeValentin.timer;
+package polytech.annecy.groupe_valentin.timer;
 
 import java.util.Iterator;
 import java.util.TreeSet;
