+++ src/main/java/polytech/annecy/groupe_valentin/timer/PeriodicTimer.java
@@ -1,4 +1,4 @@
-package polytech.annecy.groupeValentin.timer;
+package polytech.annecy.groupe_valentin.timer;
 
 public class PeriodicTimer implements Timer {
