+++ src/main/java/polytech/annecy/groupe_valentin/timer/Timer.java
@@ -1,4 +1,4 @@
-package polytech.annecy.groupeValentin.timer;
+package polytech.annecy.groupe_valentin.timer;
 
 import java.util.Iterator;
